Split monolith interfaces into own files
