+++ app/src/UI/Features/Records/Activity/forms/plant/interfaces/FormSchema.ts
@@ -1,0 +1,5 @@
+import { BaseForm, TerrestrialPlantObservationSchema, AquaticPlantObservationSchema } from '.';
+
+type FormSchema = BaseForm | TerrestrialPlantObservationSchema | AquaticPlantObservationSchema;
+
+export type { FormSchema };
