+++ app/src/UI/Features/Records/Activity/forms/plant/interfaces/index.ts
@@ -1,0 +1,4 @@
+export type { BaseForm as BaseForm } from './BaseForm';
+export type { TerrestrialPlantObservationSchema as TerrestrialPlantObservationSchema } from './TerrestrialObservation';
+export type { AquaticPlantObservationSchema as AquaticPlantObservationSchema } from './AquaticObservation';
+export type { FormSchema as FormSchema } from './FormSchema';
